Feature: Add filter support files on UI - reworks
