+++ packages/strapi-plugin-upload/admin/src/containers/SettingsPage/Wrapper.js
@@ -27,4 +27,48 @@ const Wrapper = styled.div`
   }
 `;
 
+// copy from packages/strapi-admin/admin/src/components/Webhooks/HeadersInput/index.js:28
+const getBorderColor = ({ isFocused = false }) => (isFocused ? '#78caff' : '#E3E9F3');
+
+const customStyles = {
+  control: (base, state) => ({
+    ...base,
+    border: `1px solid ${getBorderColor({ isFocused: state.isFocused })} !important`,
+    borderRadius: '2px !important',
+  }),
+  menu: base => {
+    return {
+      ...base,
+      padding: '0',
+      border: '1px solid #e3e9f3',
+      borderTop: '1px solid #78caff',
+      borderTopRightRadius: '0',
+      borderTopLeftRadius: '0',
+      borderBottomRightRadius: '3px',
+      borderBottomLeftRadius: '3px',
+      boxShadow: 'none',
+      marginTop: '-1px;',
+    };
+  },
+  menuList: base => ({
+    ...base,
+    maxHeight: '224px',
+    paddingTop: '0',
+  }),
+  option: (base, state) => {
+    return {
+      ...base,
+      backgroundColor: state.isSelected || state.isFocused ? '#f6f6f6' : '#fff',
+      color: '#000000',
+      fontSize: '13px',
+      fontWeight: state.isSelected ? '600' : '400',
+      cursor: state.isFocused ? 'pointer' : 'initial',
+      height: '32px',
+      lineHeight: '16px',
+    };
+  },
+};
+
+export { customStyles };
+
 export default Wrapper;
