+++ packages/strapi-plugin-upload/admin/src/containers/SettingsPage/index.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useReducer, useRef } from 'react';
-import CreatableSelect from 'react-select/creatable';
+import Select from 'react-select';
 import { Header, Inputs } from '@buffetjs/custom';
 import { isEqual } from 'lodash';
 import { LoadingIndicatorPage, useGlobalContext, request } from 'strapi-helper-plugin';
@@ -8,50 +8,9 @@ import { getRequestUrl, getTrad } from '../../utils';
 import Text from '../../components/Text';
 import Divider from './Divider';
 import SectionTitleWrapper from './SectionTitleWrapper';
-import Wrapper from './Wrapper';
+import Wrapper, { customStyles } from './Wrapper';
 import init from './init';
 import reducer, { initialState } from './reducer';
-
-// copy from packages/strapi-admin/admin/src/components/Webhooks/HeadersInput/index.js:28
-const getBorderColor = ({ isFocused = false }) => (isFocused ? '#78caff' : '#E3E9F3');
-const customStyles = {
-  control: (base, state) => ({
-    ...base,
-    border: `1px solid ${getBorderColor({ isFocused: state.isFocused })} !important`,
-    borderRadius: '2px !important',
-  }),
-  menu: base => {
-    return {
-      ...base,
-      padding: '0',
-      border: '1px solid #e3e9f3',
-      borderTop: '1px solid #78caff',
-      borderTopRightRadius: '0',
-      borderTopLeftRadius: '0',
-      borderBottomRightRadius: '3px',
-      borderBottomLeftRadius: '3px',
-      boxShadow: 'none',
-      marginTop: '-1px;',
-    };
-  },
-  menuList: base => ({
-    ...base,
-    maxHeight: '224px',
-    paddingTop: '0',
-  }),
-  option: (base, state) => {
-    return {
-      ...base,
-      backgroundColor: state.isSelected || state.isFocused ? '#f6f6f6' : '#fff',
-      color: '#000000',
-      fontSize: '13px',
-      fontWeight: state.isSelected ? '600' : '400',
-      cursor: state.isFocused ? 'pointer' : 'initial',
-      height: '32px',
-      lineHeight: '16px',
-    };
-  },
-};
 
 const SettingsPage = () => {
   const { formatMessage } = useGlobalContext();
@@ -71,10 +30,14 @@ const SettingsPage = () => {
           type: 'GET_DATA_SUCCEEDED',
           data: {
             ...data,
-            supportFormat: (data.supportFormat || []).map(label => ({ label, value: label })),
-            supportFormatOptions: (data.supportFormatOptions || []).map(label => ({
+            supportFormat: (data.supportFormat || []).map(label => {
+              const value = data.supportFormatOptions.find(_ => _.label === label)?.regex;
+
+              return { label, value };
+            }),
+            supportFormatOptions: (data.supportFormatOptions || []).map(({ label, regex }) => ({
               label,
-              value: label,
+              value: regex,
             })),
           },
         });
@@ -99,10 +62,11 @@ const SettingsPage = () => {
     const supportFormatOptions = modifiedData.supportFormatOptions || [];
     const payload = {
       ...modifiedData,
-      supportFormat: supportFormat.map(({ value }) => value),
-      supportFormatOptions: [
-        ...new Set(supportFormatOptions.concat(supportFormat).map(({ value }) => value)),
-      ],
+      supportFormat: supportFormat.map(({ label }) => label),
+      supportFormatOptions: supportFormatOptions.map(({ label, value }) => ({
+        label,
+        regex: value,
+      })),
     };
     try {
       await request(getRequestUrl('settings'), {
@@ -191,9 +155,8 @@ const SettingsPage = () => {
                 {formatMessage({ id: getTrad('settings.form.supportFormats.label') })}
               </label>
               <section>
-                <CreatableSelect
+                <Select
                   isMulti
-                  isClearable
                   id="support-formats"
                   onChange={onChangeSupportFormat}
                   options={modifiedData.supportFormatOptions}
